% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.new.HOBO.file.R
\name{append.new.HOBO.file}
\alias{append.new.HOBO.file}
\title{Combine previous HOBO file with new HOBO file}
\usage{
append.new.HOBO.file(
  newfile,
  file2append,
  CF_init = 1,
  TM_init,
  CF_end = 1,
  TM_end
)
}
\arguments{
\item{newfile}{data frame of new HOBO dataset that need to be appended to the old one}

\item{file2append}{data frame of old HOBO dataset the new HOBO data frame will be appended to.}

\item{CF_init}{correction factor at the beginning}

\item{TM_init}{Time measurement initialization : when probes are back in the lakes}

\item{CF_end}{correction factor at the end}

\item{TM_end}{Time measurement end: when probes are taken out of the lakes}

\item{ind_init}{index of when to start correcting the measurements}

\item{ind_end}{index of when to stop correcting the measurements}
}
\description{
Takes the previously QC'd HOBO file and append the new dataset.
}
\author{
Rosalie Bruel
}
\keyword{hobo}
\keyword{planaqua}
