% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.factor.do.R
\name{correction.factor.do}
\alias{correction.factor.do}
\title{Compute dissolved oxygen correction factors}
\usage{
correction.factor.do(
  filename,
  TC_pre,
  TC_post,
  Pbaro_mbar_init,
  Pbaro_mbar_end,
  n = 4
)
}
\arguments{
\item{filename}{path to file we need to calculate the correction factors for. The file is a raw HOBO file, with temperature and oxygen.}

\item{TC_pre}{datetime of end of the "initialization" calibration (pre-deployment)}

\item{TC_post}{datetime of end of the "end of deployement" calibration (post-deployment)}

\item{Pbaro_mbar_init}{Atmospheric pressure in millibars during the "initialization" calibration (pre-deployment)}

\item{Pbaro_mbar_end}{Atmospheric pressure in millibars during the "end of deployement" calibration (post-deployment)}

\item{n}{number of observations PRIOR to TC_pre and TC_post used to compute the oxygen mean for computing correction factor. Default to n = 4.}
}
\description{
Compute dissolved oxygen correction factor pre- and post-deployment by calculating the ratio between the theoritical dissolved oxygen concentration at saturation with the measured oxygen concentration. Requires the atmospheric pressure and the time the calculation can be made for (i.e., when the sensors are taken out their bubble bath, where the oxygen saturation should be at the maximum).
}
\author{
Rosalie Bruel
}
