% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.p.R
\name{interp.p}
\alias{interp.p}
\title{Interpolate the correction factor}
\usage{
interp.p(x, ind_init, ind_end, CF_init, CF_end)
}
\arguments{
\item{x}{HOBO dataset}

\item{ind_init}{index of when to start correcting the measurements (start of deployment)}

\item{ind_end}{index of when to stop correcting the measurements (end of deployment)}

\item{CF_init}{correction factor at the beginning}

\item{CF_end}{correction factor at the end}
}
\description{
Interpolate the correction factor at each time step between the start and end of deployment.
}
\author{
Sophie Guillon
}
\keyword{hobo}
\keyword{planaqua}
