% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.recent.files.R
\name{list.recent.files}
\alias{list.recent.files}
\title{List most recent files}
\usage{
list.recent.files(path)
}
\arguments{
\item{path}{Path to folder with all the files. Format of files within the path should be "lacXX_data_compile_YYYYMM_YYYYMM.txt". lacXX gives the lac number, and YYYYMM are the start and end dates of the recording. The second date will be used to filter out the most recent file.}
}
\description{
Will find inside the folder with all the files the file with the recent-most data, per lake. Name of each data file sould follow the format: "lacXX_data_compile_YYYYMM_YYYYMM.txt" or "lacXX_data_compile_YYYYMM.txt"
}
\author{
Rosalie Bruel
}
\keyword{hobo}
\keyword{planaqua}
