% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCF.all.R
\name{pCF.all}
\alias{pCF.all}
\title{Visualize correction factors at the beginning and end of deployment}
\usage{
pCF.all(
  x = NULL,
  alldat = TRUE,
  lakename = "Lake",
  orderby = "number",
  order = NULL,
  xmin = NULL,
  xmax = NULL
)
}
\arguments{
\item{x}{output dataframe from the correction.factor.do function}

\item{alldat}{logical argument, default = TRUE. Whether to include all the previous metadata file or not}

\item{lakename}{Character string indicating the name of the column with the lake name information in the x dataset, or alternatively, the column index.}

\item{orderby}{Choose one of c("number", "treatment", "quadrat", "own") to select the order of lakes. Default: "number".}

\item{order}{Default = NULL, or must be a numeric vector with 16 numbers, specified by the user to choose an order, e.g., c(7,8,1:4, 6, 5, 9:16).}

\item{xmin}{Minimum x you want to see. Format = "YYYY-MM-DD"}

\item{xmax}{Maximum x you want to see. Format = "YYYY-MM-DD"}
}
\description{
ggplot visual of correction factors, per lake. Uses
}
\author{
Rosalie Bruel
}
\keyword{hobo}
\keyword{planaqua}
