% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hobo.R
\name{read.hobo}
\alias{read.hobo}
\title{Read in a HOBO file}
\usage{
read.hobo(filename)
}
\arguments{
\item{filename}{path to HOBO dataset to read}
}
\description{
The function read in a hobo file in the format it is exported as at the CEREEP-Ecotron, i.e., as a .csv file. The columns are renamed and the date time format is converted to a date time format R understands.
}
\keyword{hobo}
\keyword{planaqua}
