% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhobo-package.R
\docType{package}
\name{rhobo-package}
\alias{rhobo}
\alias{rhobo-package}
\title{rhobo: Quality control assistant for HOBO data}
\description{
The R package allows to read in the raw HOBO files downloaded through the HOBO software and compute the dissolved oxygen correction factors. Functions are also included to help append new data to previous data files.
}
\author{
\strong{Maintainer}: Rosalie Bruel \email{rosaliebruel@gmail.com} (\href{https://orcid.org/0000-0001-8351-7539}{ORCID})

Authors:
\itemize{
  \item Sophie Guillon \email{sophie.guillon@mines-paristech.fr} (\href{https://orcid.org/0000-0002-1694-5477}{ORCID})
}

}
\keyword{internal}
