% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhobo.CF.R
\name{rhobo.CF}
\alias{rhobo.CF}
\title{Routine HOBO part 1: compute DO correction factors}
\usage{
rhobo.CF(
  path2data,
  foldername,
  TC_pre,
  TC_post,
  Pbaro_mbar_init,
  Pbaro_mbar_end,
  n = 4
)
}
\arguments{
\item{path2data}{User's path to the folder where the data are stored. Within that folder, there should be two subfolders with the raw data (Hobo_Raw) and the processed data (Hobo_Process)}

\item{foldername}{Folder name of the newest data that have been downloaded post-HOBO deployment. Folder name should be the date of the end of the deployment with the format YYYY_MM_DD, e.g., 2021_07_27}

\item{TC_pre}{datetime of end of the "initialization" calibration (pre-deployment)}

\item{TC_post}{datetime of end of the "end of deployement" calibration (post-deployment)}

\item{Pbaro_mbar_init}{Atmospheric pressure in millibars during the "initialization" calibration (pre-deployment)}

\item{Pbaro_mbar_end}{Atmospheric pressure in millibars during the "end of deployement" calibration (post-deployment)}

\item{n}{number of observations PRIOR to TC_pre and TC_post used to compute the oxygen mean for computing correction factor. Default to n = 4.}
}
\description{
First part of the QC routine: read in the raw file and compute dissolved oxygen correction factors.
}
\author{
Rosalie Bruel and Sophie Guillon
}
\keyword{hobo}
\keyword{planaqua}
