% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhobo.check.inputs.R
\name{rhobo.check.inputs}
\alias{rhobo.check.inputs}
\title{Check inputs before HOBO QC}
\usage{
rhobo.check.inputs(
  path2data = NULL,
  foldername = NULL,
  TC_pre = NULL,
  TC_post = NULL,
  TM_init = NULL,
  TM_end = NULL,
  Pbaro_mbar_init = NULL,
  Pbaro_mbar_end = NULL
)
}
\arguments{
\item{path2data}{User's path to the folder where the data are stored. Within that folder, there should be two subfolders with the raw data (Hobo_Raw) and the processed data (Hobo_Process)}

\item{TC_pre}{datetime of end of the "initialization" calibration (pre-deployment)}

\item{TC_post}{datetime of end of the "end of deployement" calibration (post-deployment)}

\item{TM_init}{Time measurement initialization : when probes are back in the lakes}

\item{TM_end}{Time measurement end: when probes are taken out of the lakes}

\item{Pbaro_mbar_init}{Atmospheric pressure in millibars during the "initialization" calibration (pre-deployment)}

\item{Pbaro_mbar_end}{Atmospheric pressure in millibars during the "end of deployement" calibration (post-deployment)}

\item{filename}{path to file we need to calculate the correction factors for. The file is a raw HOBO file, with temperature and oxygen.}
}
\description{
Take as input all the parameters and does some basic checks (chronological order for dates, files are present in the folder, pressure is in millibar)
}
\author{
Rosalie Bruel
}
\keyword{hobo}
\keyword{planaqua}
