% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhobo.treatments.R
\name{rhobo.treatments}
\alias{rhobo.treatments}
\title{Assign treatments}
\usage{
rhobo.treatments(x, lakename = NULL, orderby = "number", order = NULL)
}
\arguments{
\item{x}{Input dataset}

\item{lakename}{Character string indicating the name of the column with the lake name information in the x dataset, or alternatively, the column index.}

\item{orderby}{Choose one of c("number", "treatment", "quadrat", "own") to select the order of lakes. Default: "number".}

\item{order}{Default = NULL, or must be a numeric vector with 16 numbers, specified by the user to choose an order, e.g., c(7,8,1:4, 6, 5, 9:16).}
}
\description{
Create a new column for any data frame with lake treatments, based on lake names. Order the lakes by name (default) or treatment
}
\author{
Rosalie Bruel
}
\keyword{PLANAQUA}
\keyword{hobo}
