% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.DO.CF.R
\name{specify.DO.CF}
\alias{specify.DO.CF}
\title{Specify DO correction factor manually}
\usage{
specify.DO.CF(x, name, time, CF)
}
\arguments{
\item{x}{the dataframe with the output}

\item{name}{name of the lake as character, e.g., "lac01"}

\item{time}{"Init" or "End"}

\item{CF}{manual correction factor to enter}
}
\description{
After visual inspection, the user can decide to specify a correction factor manually.
}
\author{
Rosalie Bruel
}
\keyword{hobo}
\keyword{planaqua}
