% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.dir.R
\name{struct.dir}
\alias{struct.dir}
\title{Help showing the structure of the directory}
\usage{
struct.dir()
}
\description{
Print in the console the folder structure were all the data needed to run the rhobo routine should be.
}
\author{

}
\keyword{hobo}
\keyword{planaqua}
